+++ landingzones/rules.py
@@ -22,8 +22,7 @@ from projectroles import rules as pr_rules  # To access common predicates
 # Allow viewing user's own landing zones for the project
 rules.add_perm(
     'landingzones.view_zones_own',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -31,16 +30,14 @@ rules.add_perm(
 # Allow viewing all landing zones for the project
 rules.add_perm(
     'landingzones.view_zones_all',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate,
 )
 
 # Allow creating landing zones
 rules.add_perm(
     'landingzones.add_zones',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -48,8 +45,7 @@ rules.add_perm(
 # Allow modifying or deleting the user's own landing zones
 rules.add_perm(
     'landingzones.update_zones_own',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -57,7 +53,6 @@ rules.add_perm(
 # Allow modifying or deleting all landing zones
 rules.add_perm(
     'landingzones.update_zones_all',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate,
 )