+++ samplesheets/rules.py
@@ -22,8 +22,7 @@ from projectroles import rules as pr_rules  # To access common predicates
 # Allow viewing the sample sheet of the project
 rules.add_perm(
     'samplesheets.view_sheet',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor
     | pr_rules.is_project_guest,
@@ -32,8 +31,7 @@ rules.add_perm(
 # Allow creating, importing or modifying the project's JSON sample sheet
 rules.add_perm(
     'samplesheets.edit_sheet',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -41,8 +39,7 @@ rules.add_perm(
 # Allow exporting a JSON sample sheet from project
 rules.add_perm(
     'samplesheets.export_sheet',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -50,8 +47,7 @@ rules.add_perm(
 # Allow creating directory structure in iRODS
 rules.add_perm(
     'samplesheets.create_dirs',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )
@@ -59,8 +55,7 @@ rules.add_perm(
 # Allow deleting the project sample sheet
 rules.add_perm(
     'samplesheets.delete_sheet',
-    rules.is_superuser
-    | pr_rules.is_project_owner
+    pr_rules.is_project_owner
     | pr_rules.is_project_delegate
     | pr_rules.is_project_contributor,
 )